Moved release state radios to separate component to cut down on the length of the metadata form component
