+++ frontend/src/pages/SendMetadata/MetadataForm.jsx
@@ -6,6 +6,7 @@ import { Redirect, Link } from 'react-router-dom';
 import { alertActions } from '../../state/actions/alert';
 import { LoadingButton } from '../../sharedComponents/LoadingButton';
 import { useGetFormats } from '../../sharedComponents/hooks/GetFormats';
+import { ReleaseStateRadios } from './ReleaseStateRadios';
 
 const Wrapper = styled.div`
   flex: 1;
@@ -41,10 +42,6 @@ const Input = styled.input`
   font-size: 16px;
 `;
 
-const Radio = styled.input.attrs({ type: 'radio' })`
-  margin: 0px 5px 0px 0px;
-`;
-
 const Select = styled.select`
   flex: 0 0 2em;
   margin-bottom: 5px;
@@ -56,28 +53,6 @@ const TextArea = styled.textarea`
   resize: vertical;
   margin-bottom: 5px;
   font-size: 16px;
-`;
-
-const RadioLabel = styled.label`
-  display: flex;
-  flex-wrap: wrap;
-  align-items: center;
-  justify-content: center;
-  border-radius: 0.2em;
-  padding: 0.2em;
-  margin-bottom: 5px;
-  ${({ background }) => (background ? `background-color: ${background};` : '')}
-  ${({ border }) => (border ? `border: 1px solid ${border};` : '')}
-
-  &:not(:last-child) {
-    margin-right: 8px;
-  }
-`;
-
-const HorizontalWrapper = styled.div`
-  flex: 0;
-  display: flex;
-  flex-direction: row;
 `;
 
 export const MetadataForm = () => {
@@ -308,56 +283,7 @@ export const MetadataForm = () => {
             <b>Use the DCAT-NO Wizard.</b>
           </Link>
         </NewMetadataType>
-        <HorizontalWrapper>
-          <RadioLabel htmlFor="bluelight" background="#9999dd" border="#6666aa">
-            <Radio
-              type="radio"
-              name="releaseState"
-              value={1}
-              id="bluelight"
-              checked={releaseState === 1}
-              onChange={handleRadioChange}
-              required
-            />
-            {' Released'}
-          </RadioLabel>
-          <RadioLabel htmlFor="greenlight" background="#ccffcc" border="#00ff00">
-            <Radio
-              type="radio"
-              name="releaseState"
-              value={2}
-              id="greenlight"
-              checked={releaseState === 2}
-              onChange={handleRadioChange}
-              required
-            />
-            {' Ready for release'}
-          </RadioLabel>
-          <RadioLabel htmlFor="yellowlight" background="#ffffcc" border="#d8d800">
-            <Radio
-              type="radio"
-              name="releaseState"
-              value={3}
-              id="yellowlight"
-              checked={releaseState === 3}
-              onChange={handleRadioChange}
-              required
-            />
-            {' Needs work'}
-          </RadioLabel>
-          <RadioLabel htmlFor="redlight" background="#ffcccc" border="#ff5555">
-            <Radio
-              type="radio"
-              name="releaseState"
-              value={4}
-              id="redlight"
-              checked={releaseState === 4}
-              onChange={handleRadioChange}
-              required
-            />
-            {' Unreleasable'}
-          </RadioLabel>
-        </HorizontalWrapper>
+        <ReleaseStateRadios releaseState={releaseState} handleRadioChange={handleRadioChange} />
         <TextArea placeholder="Description" name="description" value={description} onChange={handleChange} required />
         <Select name="municipalityName" value={municipalityName} onChange={handleChange} required>
           <option value="" disabled>Municipality</option>
